Breast Cancer Wisconsin (Original) Data Set

Samples arrive periodically as Dr. Wolberg reports his clinical cases. The database therefore reflects this chronological grouping of the data.

Usage

bcw_original

Format

A data frame with 699 observations on the following 11 variables.

sample_code_number: id number

clump_thickness: 1 - 10

uniformity_of_cell_size: 1 - 10

uniformity_of_cell_shape: 1 - 10

single_epithelial_cell_size: 1 - 10

bare_nuclei: 1 - 10

bland_chromatin: 1 - 10

normal_nucleoli: 1 - 10

mitoses: 1 - 10

class: 2 for benign, 4 for malignant

Source

Dr. William H. Wolberg - Physician University of Wisconsin Hospitals Madison, Wisconsin, USA

Details

This grouping information appears immediately below, having been removed from the data itself:

Group	Instances	Date of Collection

1	367	January 1989

2	70	October 1989

3	31	February 1990

4	17	April 1990

5	48	August 1990

6	49	Updated January 1991

7	31	June 1991

8	86	November 1991

Total	699 points	15 July 1992

Note that the results summarized above in Past Usage refer to a dataset of size 369, while Group 1 has only 367 instances. This is because it originally contained 369 instances; 2 were removed.

References

https://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/breast-cancer-wisconsin.data