Glass Identification Data Set

The study of classification of types of glass was motivated by criminological investigation. At the scene of the crime, the glass left can be used as evidence...if it is correctly identified!

Usage

glass

Format

A data frame with 214 observations on the following 11 variables.

ID: 1 to 214

RI: refractive index

Na: weight percent in corresponding oxide, as are attributes 4-10)

Mg: Magnesium

Al: Aluminum

Si: Silicon

K: Potassium

Ca: Calcium

Ba: Barium

Fe: Iron

Type: Class attribute

1: building_windows_float_processed

2: building_windows_non_float_processed

3: vehicle_windows_float_processed

4: vehicle_windows_non_float_processed (none in this database)

5: containers

6: tableware

7: headlamps

Source

B. German Central Research Establishment Home Office Forensic Science Service Aldermaston, Reading, Berkshire RG7 4PN

Vina Spiehler, Ph.D., DABFT Diagnostic Products Corporation (213) 776-0180 (ext 3014)

References

https://archive.ics.uci.edu/ml/machine-learning-databases/glass/glass.data https://archive.ics.uci.edu/ml/datasets/Glass+Identification