Bike Sharing (Daily) Data Set

Bike sharing systems are new generation of traditional bike rentals where whole process from membership, rental and return back has become automatic. Through these systems, user is able to easily rent a bike from a particular position and return back at another position. Currently, there are about over 500 bike-sharing programs around the world which is composed of over 500 thousands bicycles. Today, there exists great interest in these systems due to their important role in traffic, environmental and health issues.

Usage

bike_sharing_daily

Format

A data frame with 731 observations on the following 16 variables.

instant: Record index

dteday: Date

season:

1: Spring

2: Summer

3: Fall

4: Winter

yr:

0: 2011

1: 2012

mnth:

1: Jan

2: Feb

3: Mar

4: Apr

5: May

6: Jun

7: Jul

8: Aug

9: Sep

10: Oct

11: Nov

12: Dec

hr:

0: 12 AM

1: 1 AM

2: 2 AM

3: 3 AM

4: 4 AM

5: 5 AM

6: 6 AM

7: 7 AM

8: 8 AM

9: 9 AM

10: 10 AM

11: 11 AM

12: 12 PM

13: 1 PM

14: 2 PM

15: 3 PM

16: 4 PM

17: 5 PM

18: 6 PM

19: 7 PM

20: 8 PM

21: 9 PM

22: 10 PM

23: 11 PM

holiday:

Whether the day is a holiday or not according to the Human Resources page of DC.

0: No

1: Yes

weekday:

The day of a week

0: Sunday

1: Monday

2: Tuesday

3: Wednesday

4: Thursday

5: Friday

6: Saturday

workingday:

Whether the day is a workday (Monday - Friday)

0: No

1: Yes

weathersit:

1: Clear, Few clouds, Partly cloudy, Partly cloudy

2: Mist + Cloudy, Mist + Broken clouds, Mist + Few clouds, Mist

3: Light Snow, Light Rain + Thunderstorm + Scattered clouds, Light Rain + Scattered clouds

4: Heavy Rain + Ice Pallets + Thunderstorm + Mist, Snow + Fog

temp:

Normalized temperature in Celsius.

The values are derived via \(\frac{(t-t_{min})}{(t_{max}-t_{min})}\), t_min=-8, t_max=+39

atemp:

Normalized feeling temperature in Celsius.

The values are derived via \(\frac{(t-t_{min})}{(t_{max}-t_{min})}\), t_min=-16, t_max=+50

hum:

Normalized humidity.

The values are divided to 100 (max)

windspeed:

Normalized wind speed.

The values are divided to 67 (max)

casual:

Count of casual users

registered:

Count of registered users

cnt:

Count of total rental bikes including both casual and registered

Source

Hadi Fanaee-T

Laboratory of Artificial Intelligence and Decision Support (LIAAD), University of Porto

INESC Porto, Campus da FEUP

Rua Dr. Roberto Frias, 378

4200 - 465 Porto, Portugal

Details

Apart from interesting real world applications of bike sharing systems, the characteristics of data being generated by these systems make them attractive for the research. Opposed to other transport services such as bus or subway, the duration of travel, departure and arrival position is explicitly recorded in these systems. This feature turns bike sharing system into a virtual sensor network that can be used for sensing mobility in the city. Hence, it is expected that most of important events in the city could be detected via monitoring these data.

References

Original Source: http://capitalbikeshare.com/system-data

Weather Information: http://www.freemeteo.com

Holiday Schedule: http://dchr.dc.gov/page/holiday-schedule