Heart Disease Processed Data Sets

Detecting the presence of heart disease in patients.

Usage

heart_disease_cl

heart_disease_hu

heart_disease_va

heart_disease_ch

Format

Four data.frames with a varying number of observations that contain the following 14 variables.

age: age in years

sex: sex (1 = male; 0 = female)

cp: chest pain type

Value 1: typical angina

Value 2: atypical angina

Value 3: non-anginal pain

Value 4: asymptomatic

trestbps: resting blood pressure (in mm Hg on admission to the hospital)

chol: serum cholestoral in mg/dl

fbs: fasting blood sugar > 120 mg/dl (1 = true; 0 = false)

restecg: resting electrocardiographic results

Value 0: normal

Value 1: having ST-T wave abnormality (T wave inversions and/or ST elevation or depression of > 0.05 mV)

Value 2: showing probable or definite left ventricular hypertrophy by Estes' criteria

thalach: maximum heart rate achieved

exang: exercise induced angina (1 = yes; 0 = no)

oldpeak: ST depression induced by exercise relative to rest

slope: the slope of the peak exercise ST segment

Value 1: upsloping

Value 2: flat

Value 3: downsloping

ca: number of major vessels (0-3) colored by flourosopy

thal: See below

Value 3: normal

Value 6: fixed defect

Value 7: reversable defect

num: diagnosis of heart disease (angiographic disease status)

Value 0: < 50% diameter narrowing

Value 1: > 50% diameter narrowing

An object of class data.frame with 294 rows and 14 columns.

An object of class data.frame with 200 rows and 14 columns.

An object of class data.frame with 123 rows and 14 columns.

Source

Hungarian Institute of Cardiology. Budapest: Andras Janosi, M.D.

University Hospital, Zurich, Switzerland: William Steinbrunn, M.D.

University Hospital, Basel, Switzerland: Matthias Pfisterer, M.D.

V.A. Medical Center, Long Beach and Cleveland Clinic Foundation: Robert Detrano, M.D., Ph.D.

Details

The data was collected from the four following locations:

heart_disease_cl: Cleveland Clinic Foundation

heart_disease_hu: Hungarian Institute of Cardiology, Budapest

heart_disease_va: V.A. Medical Center, Long Beach, CA

heart_disease_ch: University Hospital, Zurich, Switzerland

Database	Instances	Cleveland

303	Hungarian	294

Switzerland	123	Long Beach VA

References

https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.cleveland.data https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.hungarian.data https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.switzerland.data https://archive.ics.uci.edu/ml/machine-learning-databases/heart-disease/processed.va.data https://archive.ics.uci.edu/ml/datasets/Heart+Disease